# Faster Vert.x entry