--- a/vertx/src/main/java/hello/Server.java
+++ b/vertx/src/main/java/hello/Server.java
@@ -3,14 +3,16 @@ package hello;
 import io.vertx.core.AbstractVerticle;
 import io.vertx.core.DeploymentOptions;
 import io.vertx.core.Vertx;
+import io.vertx.core.buffer.Buffer;
 import io.vertx.core.http.HttpHeaders;
-
-import java.util.Date;
 
 public class Server extends AbstractVerticle {
 
-    private static final int INSTANCES = 4;
-    private static final String TEXT = "Hello World!";
+    private static final int INSTANCES = Runtime.getRuntime().availableProcessors() * 2;
+    private static final Buffer TEXT = Buffer.buffer("Hello World!");
+
+    private static final String TEXT_PLAIN = "text/plain";
+    private static final String VERTX = "Vertx";
 
     // Convenience method so you can run it in your IDE
     public static void main(String[] args) {
@@ -22,8 +24,8 @@ public class Server extends AbstractVerticle {
     public void start() throws Exception {
         vertx.createHttpServer().requestHandler(req -> {
             req.response()
-                .putHeader(HttpHeaders.SERVER, "Vertx")
-                .putHeader(HttpHeaders.CONTENT_TYPE, "text/plain")
+                .putHeader(HttpHeaders.SERVER, VERTX)
+                .putHeader(HttpHeaders.CONTENT_TYPE, TEXT_PLAIN)
                 .end(TEXT);
         }).listen(8080);
     }