1. using more verticals. since the test is done with more than 4 connections
2. all strings to constants
3. using Vert.x buffer over a string is much more efficient