@stevehu 2019-02-11
Good performance compare of both. 

old
```
wrk -t4 -c128 -d30s http://localhost:8080 -s pipeline.lua --latency -- / 16
Running 30s test @ http://localhost:8080
  4 threads and 128 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   814.14us    3.39ms 114.31ms   97.39%
    Req/Sec   471.59k    51.36k  607.19k    90.75%
  Latency Distribution
     50%  615.00us
     75%    0.88ms
     90%    1.10ms
     99%    0.00us
  56297104 requests in 30.00s, 4.82GB read
Requests/sec: 1876537.98
Transfer/sec:    164.64MB
```

new 

```
wrk -t4 -c128 -d30s http://localhost:8080 -s pipeline.lua --latency -- / 16
Running 30s test @ http://localhost:8080
  4 threads and 128 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.34ms    5.61ms 170.00ms   94.85%
    Req/Sec     0.95M   133.67k    1.23M    92.76%
  Latency Distribution
     50%  278.00us
     75%  517.00us
     90%    3.23ms
     99%  126.69ms
  113247792 requests in 30.10s, 9.70GB read
Requests/sec: 3762376.24
Transfer/sec:    330.10MB

```