@stevehu 2019-02-11
@lior-k I did a retest on light-4j and here is the result on the same computer. 

```
wrk -t4 -c128 -d30s http://localhost:8080 -s pipeline.lua --latency -- / 16
Running 30s test @ http://localhost:8080
  4 threads and 128 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   774.30us    1.77ms  40.31ms   91.51%
    Req/Sec     0.93M    81.46k    1.06M    87.83%
  Latency Distribution
     50%  286.00us
     75%  529.00us
     90%    1.92ms
     99%    0.00us
  110603088 requests in 30.01s, 12.88GB read
Requests/sec: 3685234.33
Transfer/sec:    439.31MB
```
And you can see that the transfer/sec is much higher in light-4j 439MB/s vs vertx 330MB/s as we are putting date and server name into the response header. I think other frameworks are doing the same. I am wondering if you could make some change to the vertex implementation. Thanks.